refactor users.py
