+++ main.py
@@ -7,6 +7,7 @@ from fastapi_limiter import FastAPILimiter
 from fastapi_limiter.depends import RateLimiter
 # pylint: disable=C0103, W0613, W0621, W0603
 from src.conf.config import settings
+from src.routes import auth
 from src.routes import users
 from src.routes import posts
 from src.routes import comments
@@ -50,6 +51,7 @@ async def lifespan(app: FastAPI):
 
 app = FastAPI(title="PixnTalk", lifespan=lifespan)
 
+app.include_router(auth.router, prefix='/api')
 app.include_router(users.router, prefix='/api')
 app.include_router(posts.router, prefix='/api')
 app.include_router(comments.router, prefix='/api')
